git-svn-id: http://svn.l2jfree.com/l2jfree/branches/genesis@9322 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ l2jfree-core/src/main/java/com/l2jfree/gameserver/network/client/packets/receivable/RequestMagicSkillList.java
@@ -67,7 +67,7 @@ public abstract class RequestMagicSkillList extends L2ClientPacket
 	{
 		// TODO: implement
 		L2Player activeChar = getClient().getActiveChar();
-		if (_characterId != activeChar.getCharacterId().intValue() || _objectId != activeChar.getObjectId())
+		if (_characterId != activeChar.getCharacterId().intValue() || _objectId != activeChar.getObjectId().intValue())
 			getClient().closeNow();
 	}
 }
